riscv: vector: Fix the boot issue compiled using xuantie-toolchain or upstream-toolchain
The Image compiled by upstream-toolchain gcc-14 will generate vector
illegal instruction exception when booting, because the compilation
option -march contains v, which is due to the requirement of the
lower version of xuantie-toolchain to explicitly specify v. In the
future, as xuantie-toolchain is upgraded, it will remain consistent
with upstream-toolchain and does not need to be explicitly specified.
Therefore, it is only explicitly specified for the current and old
versions of xuantie-toolchain, including the following three xuantie-
toolchain versions:
xuantie-toolchain-v2.10 with gcc-10.4.
xuantie-toolchain-v2.8 with gcc-10.4.
xuantie-toolchain-v2.6 with gcc-10.2.

Signed-off-by: Chen Pei <cp0613@linux.alibaba.com>
[ Han Gao: Resovle rebase on v6.6.94 conflicts ]
Signed-off-by: Han Gao <rabenda.cn@gmail.com>